requirements.txt
Django
djongo
djangorestframework
google-generativeai
pymongo
